Cleaned up the code

+++ src/main/java/com/mskiba/archetype/model/AbstractModel.java
@@ -2,19 +2,20 @@ package com.mskiba.archetype.model;
 
 import javax.persistence.*;
 
+
 @MappedSuperclass
 public abstract class AbstractModel {
 
     @Id
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 }
